+++ classes/bank/jazzquiz_question_bank_view.php
@@ -54,8 +54,28 @@ class jazzquiz_question_bank_view extends \core_question\local\bank\view {
             'core_question\\local\\bank\\checkbox_column',
             'qbank_viewquestiontype\\question_type_column',
             'qbank_viewquestionname\\viewquestionname_column_helper',
-            'qbank_previewquestion\\preview_action_column'
+            'qbank_previewquestion\\preview_action_column',
         ];
+
+        // Needs to check qbshowtext parameter manually from baseurl in order
+        // to determine whether to display the question text row below questions
+        $queryparams = explode(';', $this->baseurl);
+        
+        foreach ($queryparams as $value) {
+            $value = str_replace('&', '', $value);
+            $value = str_replace('amp', '', $value);
+            $param = explode('=', $value);
+
+            if ($param[0] == 'qbshowtext') {
+                if ((int)$param[1] != 0) {
+                    array_push($columns, 'qbank_viewquestiontext\\question_text_row');
+                }
+                break;
+            }
+        } 
+
+
+
         foreach ($columns as $column) {
             $this->requiredcolumns[$column] = new $column($this);
         }
@@ -100,7 +120,7 @@ class jazzquiz_question_bank_view extends \core_question\local\bank\view {
             $cat,
             null,
             $page,
-            $perpage,   
+            $perpage,
             $this->contexts->having_cap('moodle/question:add')
         );
         $this->display_add_selected_questions_button();
@@ -109,7 +129,13 @@ class jazzquiz_question_bank_view extends \core_question\local\bank\view {
 
     private function display_add_selected_questions_button() {
         $straddtoquiz = get_string('add_to_quiz', 'jazzquiz');
-        echo '<button class="btn btn-secondary jazzquiz-add-selected-questions">' . $straddtoquiz . '</button>';
+        echo 
+            '<br/>
+            <button class="btn btn-secondary jazzquiz-add-selected-questions">' . $straddtoquiz . '</button>
+            <i class="icon fa fa-question-circle text-info fa-fw " 
+                title="Add selected questions to the quiz" 
+                role="img" aria-label="Add selected questions to the quiz">
+            </i>';
     }
 
     /**